requirements.txt
boto3==1.20.63
psycopg2-binary==2.9.1